ALMOST PROD READY
We now account for an admin attempting to map a submission to another HeyPub user.  Basically, if we see meta data for the existing username and it does not map the metadata coming from heypublisher (primarily the email and oid) then we disallow.  Otherwise we create the user account locally in the WP database and associate our metadata with that user account.

We need to thoroughly test all use cases for submissions to ensure we've blocked all of the holes.

Then it will be ready for release.